+++ CHANGELOG.md
@@ -2,11 +2,10 @@
 
 ## v3.4.1
 
+- Grok: Fix missing trigger element to expand sidebar & focus management
 - Claude: adapt selector for 'recent chats' element in sidebar or focus the first button or link
 
 ### On Progress
-
--
 
 =======